\name{get.dataset.factors}
\alias{get.dataset.factors}
\title{Put factors into a choiceItem}
\usage{get.dataset.factors(item, ..., user.data=list(include.primary=FALSE))}
\description{Fill choiceItem widget with the linked factor fields of an objectItem}
\arguments{
\item{item}{Character name of widget to call function from}
\item{...}{Character names of widgets to fill with column metadata fields}
\item{user.data}{Whether to include the primary column ID in the fields or not}
}