\name{get.dataset.row.metadata.fields}
\alias{get.dataset.row.metadata.fields}
\title{Put row metadata fields into a choiceItem}
\usage{get.dataset.row.metadata.fields(item, ..., user.data=list(include.primary=FALSE))}
\description{Fill choiceItem widget with the linked row metadata fields of an objectItem}
\arguments{
\item{item}{Character name of widget to call function from}
\item{...}{Character names of widgets to fill with row metadata fields}
\item{user.data}{Whether to include the primary row ID in the fields or not}
}